refactor: move dds handling to module, cleanup iw4 loader
